Feature - rename w/ '.LogicApps' sub namespace
