+++ src/Invictus.Testing.LogicApps/Invictus.Testing.LogicApps.csproj
