+++ src/Invictus.Testing.LogicApps/LogicAppNotUpdatedException.cs
@@ -2,35 +2,35 @@
 using System.Runtime.Serialization;
 using System.Security.Permissions;
 
-namespace Invictus.Testing
+namespace Invictus.Testing.LogicApps
 {
     /// <summary>
-    /// Exception thrown when no trigger can be found for a given logic app.
+    /// Thrown when the logic app running in Azure could not be updated.
     /// </summary>
     [Serializable]
-    public class LogicAppTriggerNotFoundException : LogicAppException
+    public class LogicAppNotUpdatedException : LogicAppException
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="LogicAppTriggerNotFoundException"/> class.
+        /// Initializes a new instance of the <see cref="LogicAppNotUpdatedException"/> class.
         /// </summary>
-        public LogicAppTriggerNotFoundException()
+        public LogicAppNotUpdatedException()
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="LogicAppTriggerNotFoundException"/> class.
+        /// Initializes a new instance of the <see cref="LogicAppNotUpdatedException"/> class.
         /// </summary>
         /// <param name="message">The message that describes the exception.</param>
-        public LogicAppTriggerNotFoundException(string message) : base(message)
+        public LogicAppNotUpdatedException(string message) : base(message)
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="LogicAppTriggerNotFoundException"/> class.
+        /// Initializes a new instance of the <see cref="LogicAppNotUpdatedException"/> class.
         /// </summary>
         /// <param name="message">The message that describes the exception.</param>
         /// <param name="innerException">The exception that is the cause of the current exception</param>
-        public LogicAppTriggerNotFoundException(string message, Exception innerException) : base(message, innerException)
+        public LogicAppNotUpdatedException(string message, Exception innerException) : base(message, innerException)
         {
         }
 
@@ -41,7 +41,7 @@ namespace Invictus.Testing
         /// <param name="resourceGroup">The resource group where the logic app is located.</param>
         /// <param name="logicAppName">The name of the logic app resource running in Azure.</param>
         /// <param name="message">The message that describes the exception.</param>
-        public LogicAppTriggerNotFoundException(
+        public LogicAppNotUpdatedException(
             string subscriptionId,
             string resourceGroup,
             string logicAppName,
@@ -57,7 +57,7 @@ namespace Invictus.Testing
         /// <param name="logicAppName">The name of the logic app resource running in Azure.</param>
         /// <param name="message">The message that describes the exception.</param>
         /// <param name="innerException">The exception that is the cause of the current exception</param>
-        public LogicAppTriggerNotFoundException(
+        public LogicAppNotUpdatedException(
             string subscriptionId,
             string resourceGroup,
             string logicAppName,
@@ -70,7 +70,7 @@ namespace Invictus.Testing
         /// Initializes a new instance of the <see cref="LogicAppException"/> class.
         /// </summary>
         [SecurityPermission(SecurityAction.Demand, SerializationFormatter = true)]
-        protected LogicAppTriggerNotFoundException(SerializationInfo info, StreamingContext context) : base(info, context)
+        protected LogicAppNotUpdatedException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
     }
