+++ src/Invictus.Testing.LogicApps/Model/LogicAppTrigger.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using GuardNet;
 
-namespace Invictus.Testing.Model
+namespace Invictus.Testing.LogicApps.Model
 {
     /// <summary>
     /// Represents the trigger on an Logic App running on Azure.
