+++ src/Invictus.Testing.LogicApps/Serialization/StaticResultDefinition.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Invictus.Testing.Serialization
+namespace Invictus.Testing.LogicApps.Serialization
 {
     public class StaticResultDefinition
     {
