+++ src/Invictus.Testing.Tests.Integration/LogicAppConverterTests.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using Bogus;
 using Bogus.Extensions;
-using Invictus.Testing.Model;
+using Invictus.Testing.LogicApps;
+using Invictus.Testing.LogicApps.Model;
 using Microsoft.Azure.Management.Logic.Models;
 using Newtonsoft.Json;
 using Xunit;
