+++ src/Invictus.Testing.Tests.Integration/LogicAppsProviderTests.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Invictus.Testing.Model;
+using Invictus.Testing.LogicApps;
+using Invictus.Testing.LogicApps.Model;
 using Xunit;
 using Xunit.Abstractions;
