+++ src/Invictus.Testing.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invictus.Testing", "Invictus.Testing\Invictus.Testing.csproj", "{43477A7C-3294-4E0F-A29C-03D5758BAF17}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Invictus.Testing.LogicApps", "Invictus.Testing.LogicApps\Invictus.Testing.LogicApps.csproj", "{43477A7C-3294-4E0F-A29C-03D5758BAF17}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{65BD93B8-385C-4C14-9773-140D5058D6B1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invictus.Testing.Tests.Integration", "Invictus.Testing.Tests.Integration\Invictus.Testing.Tests.Integration.csproj", "{621838F3-EF62-46C4-9517-77C2FF919A29}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Invictus.Testing.Tests.Integration", "Invictus.Testing.Tests.Integration\Invictus.Testing.Tests.Integration.csproj", "{621838F3-EF62-46C4-9517-77C2FF919A29}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
